requirements.txt
fastapi
uvicorn
scikit-learn
sentence-transformers
numpy
pandas
faiss-cpu
scikit-fuzzy
jupyter